organize all related components to Settings page to the same package.
